Made Sprites
Im not sure which branch this belongs in.... Putting it in master. Amost
all sprites are 36x36 pixels as in the original game. The only sprites
which are a different size are the Tulip_Eating_X sprites, because they
must span two tiles. All sprites are .png images and are transparent.